setup and navigate to restaurant detail

+++ JData/Sources/JData/UseCases/Restaurant/Dish.swift
@@ -5,17 +5,17 @@
 import Foundation
 
 public struct Dish: Identifiable {
-		public let id: String
-		public let name: String
-		public let price: Double
-		public let description: String
-		
-		public init(id: String, name: String, price: Double, description: String) {
-				self.id = id
-				self.name = name
-				self.price = price
-				self.description = description
-		}
+	public let id: String
+	public let name: String
+	public let price: Decimal
+	public let description: String
+	
+	public init(id: String, name: String, price: Decimal, description: String) {
+		self.id = id
+		self.name = name
+		self.price = price
+		self.description = description
+	}
 }
 
 extension Dish: Codable {}
